+++ src/llmq/blockprocessor.cpp
@@ -459,7 +459,7 @@ std::vector<const CBlockIndex*> CQuorumBlockProcessor::GetMinedCommitmentsUntilB
         }
 
         uint32_t nMinedHeight = std::numeric_limits<uint32_t>::max() - be32toh(std::get<2>(curKey));
-        if (nMinedHeight > pindex->nHeight) {
+        if (nMinedHeight > uint32_t(pindex->nHeight)) {
             break;
         }
